ASAP-97 refactor: Bootstrap/common 패키지 구조 변경
