+++ Bootstrap-Module/src/main/kotlin/com/asap/bootstrap/common/config/WebConfig.kt
@@ -1,4 +1,4 @@
-package com.asap.bootstrap.common
+package com.asap.bootstrap.common.config
 
 import org.springframework.context.annotation.Configuration
 import org.springframework.web.servlet.config.annotation.CorsRegistry
